Part 1 of PR feedback, filename changes
